+++ changes.md
@@ -3,12 +3,16 @@
 ## Main wip
 
 ## 3.2.0
-* Added assembly build to shieldwall straight.
+* Added assembly build for shieldwall components.
+  * Straight
+  * EndCap
 * Fix ShieldShape bug where the base height would break in a way that made the shape not actually symetrical.
 * Shieldwall set added support for setting base_height to all parts.
 * Shieldwall added Magnets
 * Integrated Magnets into 
   * StraightBasic
+  * Straight
+  * EndCap
 
 ## 3.1.0
 * Added shieldwall Set